requirements.txt
# Required dependencies
openai
requests
pyyaml

# Required for document conversion
markitdown